requirements.txt
yt-dlp>=2025.01.12
openai-whisper==20240930
numpy<2
